add "ExcludesPaths" middleware trait
