+++ src/Illuminate/Foundation/Http/Middleware/Concerns/ExcludesPaths.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace Illuminate\Foundation\Http\Middleware\Concerns;
+
+trait ExcludesPaths
+{
+    /**
+     * Get the URIs that should be accessible even when maintenance mode is enabled.
+     *
+     * @return array
+     */
+    public function getExcludedPaths()
+    {
+        return $this->except;
+    }
+
+    /**
+     * Determine if the request has a URI that should be accessible in maintenance mode.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return bool
+     */
+    protected function inExceptArray($request)
+    {
+        foreach ($this->getExcludedPaths() as $except) {
+            if ($except !== '/') {
+                $except = trim($except, '/');
+            }
+
+            if ($request->fullUrlIs($except) || $request->is($except)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+}
